Add relations into swagger yaml file
