+++ app/controllers/admins_controller.rb
@@ -9,13 +9,8 @@ Generated by: https://github.com/swagger-api/swagger-codegen.git
 
 =end
 class AdminsController < ApplicationController
-
-  def add_action
-    stage = Stage.find(params[:stage_id])
-    action = stage.actions.create!(action_params)
-
-    json_response(action, :created)
-  end
+  include UserOperationsMixin
+  include ApproverOperationsMixin
 
   def add_group
     group = Group.create!(group_params)
@@ -39,18 +34,6 @@ class AdminsController < ApplicationController
     json_response(workflow, :created)
   end
 
-  def fetch_action_by_id
-    action = Action.find(params[:id])
-
-    json_response(action)
-  end
-
-  def fetch_actions
-    actions = Action.all
-
-   json_response(actions)
-  end
-
   def fetch_group_by_id
     group = Group.find(params[:id])
 
@@ -67,12 +50,6 @@ class AdminsController < ApplicationController
     reqs = Request.filter(params.slice(:requester, :decision, :state))
 
     json_response(reqs)
-  end
-
-  def fetch_stage_by_id
-    stage = Stage.find(params[:id])
-
-    json_response(stage)
   end
 
   def fetch_stages
@@ -114,12 +91,6 @@ class AdminsController < ApplicationController
     json_response(workflows)
   end
 
-  def remove_action
-    Action.find(params[:id]).destroy
-
-    head :no_content
-  end
-
   def remove_group
     Group.find(params[:id]).destroy
     head :no_content
@@ -145,12 +116,6 @@ class AdminsController < ApplicationController
     head :no_content
   end
 
-  def update_action
-    Action.find(params[:id]).update(action_params)
-
-    head :no_content
-  end
-
   def update_group
     Group.find(params[:id]).update(group_params)
     head :no_content
@@ -158,11 +123,6 @@ class AdminsController < ApplicationController
 
   def update_request
     RequestUpdateService.new(params[:id]).update(request_params)
-    head :no_content
-  end
-
-  def update_stage
-    StageUpdateService.new(params[:id]).update(stage_params)
     head :no_content
   end
