+++ app/controllers/approvers_controller.rb
@@ -9,46 +9,6 @@ Generated by: https://github.com/swagger-api/swagger-codegen.git
 
 =end
 class ApproversController < ApplicationController
+  include ApproverOperationsMixin
 
-  def add_action
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def fetch_action_by_id
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def fetch_actions
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def fetch_stage_by_id
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def remove_action
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def update_action
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
-
-  def update_stage
-    # Your code here
-
-    render json: {"message" => "yes, it worked"}
-  end
 end
