+++ config/application.rb
@@ -26,5 +26,7 @@ module ServiceApproval
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
+
+    config.autoload_paths << Rails.root.join("app", "controllers", "mixins").to_s
   end
 end
